Add a config option for if the cache should be purged before and after the program is run. Useful for debugging

+++ src/main/kotlin/org/abimon/spiral/core/data/CacheHandler.kt
@@ -80,7 +80,8 @@ object CacheHandler {
 
     init {
         Runtime.getRuntime().addShutdownHook(Thread {
-            cacheDir.listFiles().forEach { it.delete() }
+            if(SpiralModel.purgeCache)
+                cacheDir.listFiles().forEach { it.delete() }
         })
     }
